Home UI作成(component化)
