+++ app/_components/PostList/index.stories.ts
@@ -1,0 +1,16 @@
+import type { Meta, StoryObj } from "@storybook/react";
+
+import PostList from "./index";
+
+// More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
+const meta = {
+  title: "Example/PostList",
+  component: PostList,
+} satisfies Meta<typeof PostList>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
+  args: {},
+};
